requirements.txt
alembic==1.8.1
apturl==0.5.2
asgiref==3.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
duplicity==0.8.21
fasteners==0.14.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
future==0.18.2
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.2
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
keyring==23.5.0
kiwisolver==1.4.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
networkx==2.8.4
numpy==1.23.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.62
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.0
pipenv==2022.9.2
platformdirs==2.5.2
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.9.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
PyWavelets==1.3.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
systemd-python==234
threadpoolctl==3.1.0
tifffile==2022.5.4
typing_extensions==4.2.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.7
uvicorn==0.17.0.post1
virtualenv==20.13.0
virtualenv-clone==0.5.7
wadllib==1.3.6
Werkzeug==2.1.2
xdg==5
xkit==0.0.0
zipp==3.8.0
